// frontend/user-panel/src/user/services/swagger-generated/model/userAdressQueryResponse.ts
/**
 * UserControlPanel
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UserAdressQueryResponse { 
    cep?: string;
    street?: string;
    complement?: string;
    neighbourhood?: string;
    city?: string;
    state?: string;
    ibge?: string;
    gia?: string;
    ddd?: number;
    siafi?: string;
}